Python/requirements.txt
hidapi
torch
numpy
filterpy
scikit-learn
matplotlib
numpy-quaternion
openvr